test(e2e): D&Dの複雑なハックロジックを DropZoneArea POM にカプセル化
